fixed problems with api level 9 (#26)
